refactor(wt-1421): inline TileMenuButton into call_tile.dart
